moved test class
